fix: 당첨 번호를 두자리 수 이상으로 입력할 경우 형변환이 되지 않아 예외 처리가 발생하는 오류로 형변환 수정
Integer에서 Long으로 변경

+++ src/main/java/lotto/domain/LottoNumber.java
@@ -61,7 +61,7 @@ public class LottoNumber {
     private void validateOnlyNumberAndComma(String value) {
         String removeComma = value.replace(",","");
         try {
-            Integer.parseInt(removeComma);
+            Long.parseLong(removeComma);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 당첨 번호는 공백 없이 쉼표(,)로 구분된 숫자로만 입력해주세요.");
         }
